added single-speaker flag to allow multithreading

+++ gen_corpus_acoustic_model.py
@@ -74,8 +74,8 @@ def main():
         utt+=1
         filename = os.path.join(cur_path,'corpus',filename)
         ### EXTRACT WAV ###
-        wavfile.write(filename,sr,waveform.astype(np.int16))
-        #wavfile.write(filename,sr,waveform)
+        #wavfile.write(filename,sr,waveform.astype(np.int16))
+        wavfile.write(filename,sr,waveform)
         ### GEN TEXTGRID ###
         raw_tg = gen_naive_textgrid(waveform,sr,row['text'])
         tg = open(filename.replace('.wav', '.TextGrid'), 'w')
